fix issue related to bailleur management on convention
